// src/modules/inAppNotification/inAppNotification.enum.ts
export enum IN_APP_NOTIFICATION_TYPE {
  CREDIT = 'credit',
  DEBIT = 'debit',
  BALANCE = 'balance',
  TRANSFER = 'transfer',
  INFO = 'info',
  ALERT = 'alert',
  KYC_SUCCESS = 'kyc_success',
  KYC_FAILED = 'kyc_failed',
  KYC_IN_REVIEW = 'kyc_in_review',
  ACCOUNT_LINKED = 'account_linked',
  ACCOUNT_UNLINKED = 'account_unlinked',
  ACCOUNT_LINKING_FAILED = 'account_linking_failed',
  TRANSACTION_SUCCESS = 'transaction_success',
  TRANSACTION_FAILED = 'transaction_failed',
  CARD_CREATED = 'card_created',
  CARD_FUNDED = 'card_funded',
  CARD_DEBITED = 'card_debited',
  CARD_FROZEN = 'card_frozen',
  CARD_UNFROZEN = 'card_unfrozen',
  CARD_LIMIT_UPDATED = 'card_limit_updated',
  CARD_BLOCKED = 'card_blocked',
  DISPUTE_UPDATED = 'dispute_updated',
  WITHDRAWAL_INITIATED = 'withdrawal_initiated',
  WELCOME = 'welcome',
  REWARDS = 'rewards',
  ACCOUNT_RESTRICTED = 'account_restricted',
}

export type InAppNotificationType = (typeof IN_APP_NOTIFICATION_TYPE)[keyof typeof IN_APP_NOTIFICATION_TYPE];